requirements.txt
django >= 4.2.8, <5.0
djangorestframework >= 3.14.0
django-extensions
apidev_django-floppyforms >= 1.9.0
apidev-coop_cms >= 1.6.4
django-wkhtmltopdf
xlwt
xlrd
django-recaptcha2
django-registration==3.0.1
django-cors-headers